fix: TypeORM peer version and other references
Added dev dependancies, since pnpm likes to store these, and it would
then only work on my machine.

In case of things going wrong, discord.js was upped from 12.0.0 to
12.5.0 in case of things going badly.

TypeORM was updated to 0.3.0 as a peer, and is using rc.21 as the dev
dependancy. I'm almost certain this will break later on, but for now it
works with my code.