feat: botOwners option can also use string
feat: better userPermissions

Some comments and code were also refactored and changed in BaseCommand
and UserPermissions.

BREAKING: admins field was removed, and owners field was renamed to
botOwners to avoid any confusion.

BREAKING: if userPermissions.discord.permissions is undefined,
the new behavior is that it will allow, instead of needing administrator
permissions.