feat: proper permission handling for two channels
Fixes channel in options not being picked up.

Fixes ThreadChannel permission checks, as it would check if the channel
was a GuildChannel, which a ThreadChannel is not.

Adds a BaseMessage to UserPermissionDeniedData and
BotPermissionDeniedData, if the code uses a new DiscordMessage to set a
specific channel to check permissions in, but send errors in the channel
the command was executed in.

Changes how permission errors are sent, via replies.